Fix: Trigger callback even if dismissed early
When the modal is dismissed (confirmed/cancelled) before the animation
is done (really, within 500ms of the modal being opened) the callback
function would not be called.

This is because line 70-76 in dev/modules/handle-click.js checks whether
the modal is "visible" -- by checking if the `visible` class is set --
and only triggers callbacks if that class has been set. As the class is
set after 500ms this leaves a 500ms window after the modal opens where
the callback will not be triggered.

From my investigation, I could not find the `visible`-class being used
for anything else than exactly these checks. It has no CSS bound to it.

Therefore the 500ms delay on setting the `visible`-class has been
removed.